fix: update comment position, content 게시글과 댓글 일대다(cascade.ALL) 관계맺음

+++ backend/src/main/java/com/oldandsea/pcb/service/CommentService.java
@@ -97,13 +97,13 @@ public class CommentService {
     }
 
     @Transactional
-    public CommentUpdatePositionResponseDTO updatePosition(List<CommentUpdatePositionList> requestDTO, Long memberId) {
+    public CommentUpdatePositionResponseDTO updatePosition(List<CommentUpdatePositionList> requestDTO, Long boardId) {
         List<CommentUpdatePositionList> responseDTOList = new ArrayList<>();
         for (CommentUpdatePositionList request : requestDTO) {
             Long commentId = request.getCommentId();
             Long after = request.getAfter();
-            Comment comment = findByCommentIdAndMemberMemberId(commentId,memberId);
-            if (after == -1|| checkAfter(after, comment.getBoard().getBoardId())) {
+            Comment comment = findByCommentIdAndBoardId(commentId,boardId);
+            if (after == -1 || checkAfter(after, comment.getBoard().getBoardId()) ) {
                 comment.updatePosition(request.getAfter(), request.getPosition());
                 if(Objects.equals(comment.getCommentId(), comment.getAfter())) {
                     throw new IllegalArgumentException("commentId랑 After랑 같을 수 없습니다");
@@ -118,17 +118,16 @@ public class CommentService {
     @Transactional
     public CommentUpdateContentResponseDTO updateContent(List<CommentUpdateContentList> requestDTO, Long memberId) {
         List<CommentUpdateContentList> responseDTOList = new ArrayList<>();
-        for(CommentUpdateContentList request : requestDTO) {
+        for (CommentUpdateContentList request : requestDTO) {
             Long commentId = request.getCommentId();
-            Comment comment = findByCommentIdAndMemberMemberId(commentId, memberId);
+            Comment comment = findByCommentIdAndMemberId(commentId, memberId);
             comment.updateColorAndContent(request.getColor(), request.getContent());
             responseDTOList.add(toUpdateContentDTO(comment));
-        }
-            return CommentUpdateContentResponseDTO.builder()
-                    .updateContentList(responseDTOList)
-                    .build();
+            }
+        return CommentUpdateContentResponseDTO.builder()
+                .updateContentList(responseDTOList)
+                .build();
     }
-
 
     private CommentUpdatePositionList toUpdatePositionDTO(Comment comment) {
         return CommentUpdatePositionList.builder()
@@ -154,14 +153,37 @@ public class CommentService {
         }else
             throw new IllegalArgumentException("after에 해당하는 Id를 가진 comment가 같은 게시글에 있지 않습니다");
     }
-    private Comment findByCommentIdAndMemberMemberId(Long commentId, Long memberId) {
-        return commentRepository.findByCommentIdAndMemberMemberId(commentId, memberId).orElseThrow(
-                () -> new IllegalArgumentException("Comment doesn't exsist")
-        );
-    }
+    private Comment findByCommentIdAndBoardId(Long commentId, Long memberId) {
+            return commentRepository.findByCommentIdAndMemberMemberId(commentId, memberId).orElseThrow(
+                    () -> new IllegalArgumentException("Comment doesn't exsist")
+            );
+        }
+
     @Transactional
     public CommentUpdatePositionResponseDTO deleteCommnet(List<CommentUpdatePositionList> requestDTO, Long memberId, Long commentId) {
-        commentRepository.deleteById(commentId);
-        return updatePosition(requestDTO,memberId);
+        if(checkCommentCreator(commentId,memberId)) {
+            commentRepository.deleteById(commentId);
+            return updatePosition(requestDTO,memberId);
+        }
+        throw new IllegalArgumentException("댓글 작성자만 댓글을 삭제 할 수 있습니다");
+    }
+
+    public Boolean checkBoardCreator(Long boardId, Long memberId) {
+        Long checkBoardCreatorId = boardRepository.findById(boardId).orElseThrow(
+                () -> new IllegalArgumentException("boardId에 해당하는 board가 존재하지 않습니다(댓글 위치 수정)")
+        ).getMember().getMemberId();
+        return checkBoardCreatorId.equals(memberId);
+    }
+
+    private Boolean checkCommentCreator(Long commentId, Long memberId) {
+        Long commentCreatorId = commentRepository.findById(commentId).orElseThrow(
+                () -> new IllegalArgumentException("commentId에 해당하는 comment가 존재하지 않습니다(댓글 내용 수정)")
+        ).getMember().getMemberId();
+        return commentCreatorId.equals(memberId);
+    }
+    private Comment findByCommentIdAndMemberId(Long commnetId, Long memberId) {
+        return commentRepository.findByCommentIdAndMemberMemberId(commnetId,memberId).orElseThrow(
+                () -> new IllegalArgumentException("댓글 작성자만 댓글 내용을 수정할 수 있습니다")
+        );
     }
 }
